feat: 장바구니 도메인 API 추가

+++ src/main/kotlin/ecommerce/server/cart/domain/model/CartDetail.kt
@@ -18,9 +18,9 @@ class CartDetail(
     val productId: Long,
 
     @Column(nullable = false)
-    var quantity: Int,
+    var quantity: Long,
 
     @Column(nullable = false)
-    var price: Int,
+    var price: Long,
 ): BaseEntity() {
 }
